renamed folders with better semantics, cleaned up unused imports, debugged generate cache
