+++ README.md
@@ -1,20 +1,20 @@
 # PokemonAPI
-*This project was created as a tutorial and template to teach my brother back end web app development and general coding practices.*
 
-## ~ A tribute to the first 151 pokemon from Bulbasaur to Mew ~
-#### The function of this Back End server is to query the poekapi.co API requesting information for the original 151. With the data response from pokeapi, the data is then filtered and assigned to pokemon models that will only contain the data about the pokemon that will be displayed in the Front End (ex: Name, Stats, Description, Image) 
+*This project was created as a tutorial and template to teach my brother back end web app development and general coding practices, as well as to reenforce my own understanding*
 
 
+## Approach
+- The function of this back end server API is to query the poekapi.co API requesting data for the original 151 pokemon from Blubasaur to Mew by tailoring the path and query string with the parameters that will specifically deliver our request.
+- The pokemon data is then filtered through a pokemon class constructor into pokemon objects, contain the specific data about the pokemon that we want displayed in the Front End (ex: Pokemon name, stats, description, image).
+- A array of pokemon names is also created that will be server for search functionality and search filtering on the front end.
+- Both constructed pokemon objects and the pokemon name array are then stored into a local server cache to reduce the need for requests to pokeAPI.
+
+#### 
 
 ### Technologies, Frameworks, Libraries, and resources used:
-- PokemonAPI : [https://pokeapi.co/](https://pokeapi.co/)
-- W3Shcools : [https://www.w3schools.com/js/default.asp](https://www.w3schools.com/js/default.asp)
-
-<br>
 
 - Visual Studio Code
 - Git Bash
-- Chrome
 
 <br>
 
@@ -24,4 +24,8 @@
 - CORS
 - dotenv
 - eslint
-- Auth0
+
+<br>
+
+- PokemonAPI : [https://pokeapi.co/](https://pokeapi.co/)
+- W3Shcools : [https://www.w3schools.com/js/default.asp](https://www.w3schools.com/js/default.asp)
