+++ modules/generateCache.js
@@ -1,0 +1,10 @@
+const axios = require('axios');
+require('dotenv').config();
+
+let generateCache = async () => 
+{
+    await axios.get(process.env.GENERATE_CACHE)
+
+}
+
+module.exports = generateCache;
