+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "pokemonapi",
   "version": "1.0.0",
-  "description": "",
+  "description": "Pokemon API querying, and data handling",
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
